introduced test suite + refactor
