+++ acc/ACC.c
@@ -9,13 +9,13 @@
 #include "ACC.h"
 
 //Globals
-float hostVel;
-float hostVelPrev; //Host velocity in the previous time step
-float integral = 0; //Needs to be global as it is used in the calculation of next step
-float errorPrev = -999.0; //Previous value of PID error
-float leadDistPrev = 0;
-float overshoot = 0;
-float leadTime = 0;
+float hostVel; // ---- acc params
+float hostVelPrev; //Host velocity in the previous time step ---- acc params
+float integral = 0; //Needs to be global as it is used in the calculation of next step ---- pid params
+float errorPrev = -999.0; //Previous value of PID error ---- pid params
+float leadDistPrev = 0;// ---- acc params
+float overshoot = 0;// ---- test/debug params 
+float leadTime = 0;// ---- test/debug params
 
 void vehDynamics(float torque){
     //torque will be a value from -1 to 1, where negative values represent braking torque and positive values represent acceleration
@@ -136,32 +136,3 @@ void leadDynamics(float* dynamics, float timeStamp, float leadDistInit, float le
     dynamics[0] = leadVel;
     dynamics[1] = leadDist;
 }
-
-int main (){
-    //Define Calibrations
-    float setSpeed = 50/3.6;
-    float hostVelInit = 50/3.6;
-    float leadVelInit = 50/3.6;
-    float leadDistInit = 1000;
-    int mode = LINEAR; //0 = Constant, 1 = Linear, 2 = Stepwise, 3 = Parabolic, 4 = Sinusoidal
-    hostVel = hostVelInit;
-
-    printf("Time| Error   | Torque  | HostVel | LeadDist| LeadVel|Mode\n");
-    int steps = simTime/timeStep;
-    for (int i = 0; i <= steps; i++){
-        printf("%04.1f| ", i*timeStep);
-
-        //Calculate the lead vehicle speed and distance
-        float leadDyn[2];
-        leadDynamics(leadDyn, i*timeStep, leadDistInit, leadVelInit, mode);
-
-        //Initial Speed, Lead Velocity, Set Speed, Lead Distance
-        accAlgo(&hostVelInit, leadDyn[0], setSpeed, leadDyn[1]);
-        
-    }
-
-    printf("\nMax overshoot: %f\nLead Time: %f",-1*overshoot, leadTime);
-    return 0;
-}
-
-//Time| Error   | Torque  | HostVel | LeadDist| LeadVel|Mode
